requirements.txt
streamlit
anthropic
google-api-python-client
sentence-transformers
faiss-cpu
